Only one DB wrapper, no more separate wrappers for each database class; Tests can be run without any commenting/uncommenting by running phpunit CLI command
