+++ README.md
@@ -1,13 +1,5 @@
 # SIMPLE APP
 
 * Implemented PHPUnit testing
-* MySQL tests:
-    * Employee model and it's properties
-    * EmployeeService for getOne(), getAll() and create() methods
-* MongoDB tests:
-    * Employee model and it's properties
-    * EmployeeService for getOne(id), getAll(), create() and delete(id) methods
-    
-<br>
-<br>
-**Before running the tests, make sure that either EmployeeServiceMongoTest OR EmployeeServiceMySQLTest is commented out** 
+    * Employee model test
+    * EmployeeService CRUD methods testing - getOne(id), getAll(), create(), delete(id)
