+++ html/index.html
@@ -1,0 +1,81 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
+    <meta charset="utf-8" />
+    <title>Webserver Landing Page</title>
+    <meta name="generator" content="PuPHPet.com" />
+    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1" />
+    <link href="//netdna.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap.min.css" rel="stylesheet" />
+    <style type="text/css">pre { background: #444; color:#fff; border-style: none; }</style>
+    <!--[if lt IE 9]><script src="//html5shim.googlecode.com/svn/trunk/html5.js?"></script><![endif]-->
+</head>
+<body>
+<div class="container">
+    <div class="row">
+        <div class="col-md-12">
+            <div class="row">
+                <div class="col-md-1"></div>
+                <div class="col-md-10">
+                    <h1 class="page-header">Congratulations! You are pretty awesome.</h1>
+
+                    <p class="lead">You are well on your way to a highly customized server.</p>
+
+                    <h3>However, if you are seeing this page, it means you are using IP Address, not virtual host!</h3>
+
+                    <p>This is probably not the page you want.</p>
+
+                    <p>
+                        This page is being served from <code>/var/www/html/index.html</code>, and is the default
+                        vhost (virtual host) that comes with most web servers.
+                    </p>
+
+                    <p>
+                        What you want to do is to add all vhosts that you defined on
+                        <a href="https://puphpet.com" target="_blank">PuPHPet.com</a> to your hosts file.
+                    </p>
+
+                    <p>
+                        On OSX and Linux, this is located at <code>/etc/hosts</code>, on Windows it is located at
+                        <code>C:\Windows\System32\drivers\etc\hosts</code>.
+                    </p>
+
+                    <p>
+                        Add a line with the IP address of this server and all vhosts you chose. For example, if the
+                        IP address is <code>192.168.100.1</code>, and you chose vhost <code>awesome.dev</code>,
+                        you will want to enter the following:
+                    </p>
+
+                    <pre>192.168.100.1 awesome.dev www.awesome.dev</pre>
+
+                    <p>
+                        If you forgot the IP address you chose, check the config file at
+                        <code>puphpet/config.yaml</code>.
+                    </p>
+
+                    <h3>Still running into problems?</h3>
+
+                    <p>
+                        First, <a href="https://github.com/puphpet/puphpet/blob/master/archive/puphpet/shell/ascii-art/important-notices.txt"
+                                  target="_blank">read the instructions here.</a>
+                        Seriously, read the instructions first!
+                    </p>
+
+                    <p>
+                        If that does not solve your issue, head over to the
+                        <a href="https://github.com/puphpet/puphpet/issues" target="_blank">PuPHPet Github issue track</a>
+                        and create a new ticket. Make sure to include the contents of your
+                        <code>puphpet/config.yaml</code> file (and make sure to indent it by 4 spaces to enable
+                        Github's code syntax) and any error messages you ran into.
+                    </p>
+
+                    <p>Good luck, and happy programming!</p>
+
+                    <p>Juan Treminio</p>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+</body>
+</html>
